introduce hfuse types to pure code
In order to get the pure code closer and closer to the types needed
by the real code, introduce the FileStat and related types from Hfuse
into the pure code.

In order to not corrupt the pure code too much, placeholder FileStat
constructions for files and directories are defined.

If currentDirectory and IncreDirectory defined accurate FileStat
structures, then the output of incrementReadDirectory would include
the correct metadata for the final files. Thus this should actually
work properly once plumbed into the Fuse mechanics :-)